test: merge coalescing tests into status_manager_test.go
